requirements.txt
streamlit>=1.32.0
groq>=0.4.0
sounddevice>=0.4.6
wavio>=0.0.8
python-dotenv>=1.0.0
numpy>=1.24.0
gtts>=2.5.0
pygame>=2.5.0
requests>=2.31.0